Move `invalid_expression` test
